refactor: refactor to utilize new BE functionality
