+++ cdk/bin/cdk.ts
@@ -1,10 +1,10 @@
 #!/usr/bin/env node
 import 'source-map-support/register';
-import * as cdk from 'aws-cdk-lib';
-import {NodejsAWSShopFrontendStack} from '../lib/cdk-stack';
+import {App} from 'aws-cdk-lib';
+import {NodejsAWSShopFrontendStack} from '../lib/frontend-stack';
 
-const app = new cdk.App();
-new NodejsAWSShopFrontendStack(app, 'CdkStack', {
+const app = new App();
+new NodejsAWSShopFrontendStack(app, 'NodejsAWSShopFrontendStack', {
     /* If you don't specify 'env', this stack will be environment-agnostic.
      * Account/Region-dependent features and context lookups will not work,
      * but a single synthesized template can be deployed anywhere. */
