+++ cdk/test/cdk.test.ts
@@ -3,7 +3,7 @@
 // import * as Cdk from '../lib/cdk-stack';
 
 // example test. To run these tests, uncomment this file along with the
-// example resource in lib/cdk-stack.ts
+// example resource in lib/frontend-stack.ts
 test('SQS Queue Created', () => {
 //   const app = new cdk.App();
 //     // WHEN
