+++ src/constants/apiPaths.ts
@@ -1,7 +1,7 @@
 const API_PATHS = {
-  product: "https://ydqjmzusye.execute-api.us-east-1.amazonaws.com/prod",
+  product: "https://d93aaj5lj9.execute-api.us-east-1.amazonaws.com/prod",
   order: "https://.execute-api.eu-west-1.amazonaws.com/dev",
-  import: "https://.execute-api.eu-west-1.amazonaws.com/dev",
+  import: "https://qdde42rit0.execute-api.us-east-1.amazonaws.com/prod",
   bff: "https://.execute-api.eu-west-1.amazonaws.com/dev",
   cart: "https://.execute-api.eu-west-1.amazonaws.com/dev",
 };
